Implemented `ITopicMappingService`
Updated `TopicController` to translate the `CurrentTopic` to a DTO using the `ITopicMappingService`, which must now be injected as a dependency via the constructor. In addition, refactored the `TopicViewResult` class to rely on either an `ITopicViewModelCore` implementation, or an explicitly provided `contentType` and `view` parameter. Removed the legacy `TopicViewModel` class, which included a reference to a `Topic` entity, and an `ITopicRepository` service, neither of which are desirable in a view model.