Updated `CodeContracts.MSBuild`
Updated to the latest version from Igor Oleinikov's NuGet package. This version seems to have some issues connecting to the cache, which will require some additional work if we want to reintroduce static checking. That said, right now Code Contracts is mostly being maintained for backward compatibility and basic guard conditions, so full functionality is not a priority.