Introduced `NavigationViewModel`
The `NavigationViewModel` provides a strongly-typed view model of `IPageTopicViewModelCore` implementations associated with a root node of a navigation tier. There may be multiple `NavigationViewModel`s for a site, for different tiers or levels of navigation.

The view model does not assume a concrete type for the underlying `NavigationRoot`; that is provided by a generic, so long as the implementation implements `IPageTopicViewModelCore`.