Introduced reusable `LayoutController`
This provides a base controller for returning `PartialViewResult`s related to different aspects of a `_layout.cshtml`. Out of the box, the `LayoutController` includes a `Menu()` action for writing the top-tier navigation. It also includes several helper methods that make it trivial to add additional actions as needed (e.g., `Footer()`, AncillaryNavigation()`, &c.).