Prevented created of duplicate children
If the `ITopicMappingService` returns an existing obect (e.g., if the object is cached, or if the mapping service followed relationships) then adding children could result in duplicates. This is a _bit_ of a leaky abstraction since implementers shouldn't need to be aware of whether the `ITopicMappingService` is caching. That said, this is a bit of an edge case since typically the `ITopicMappingService` adds children, so the extra level of caution is warranted regardless.