Configured `TopicFactory` to use `DefaultTopicLookupService`
Previously, the `TopicFactory` defaulted to using the `DynamicTypeLookupService`. This would have been valuable when derived Topics were the primary point of extensibility within OnTopic. With the `ITopicMappingService`, however, there is rarely a need to extend `Topic` outside of the core library (which is accounted for by `DefaultTopicLookupService`). The `DynamicTopicLookupService` is still available for applications that require it, but it is no longer the default implementation. Customers with one or two custom types are advised to simply extend `DefaultTopicLookupService`. Customers with extensive use of derived types should configure `TopicFactory` to use `DynamicTopicLookupService`.