Moved `GetContentTypeDescriptors()` to `TopicRepositoryBase`
Generally, `GetContentTypeDescriptors()` should be able to operate independent of any underlying implementation; it is a convenience method to consolidate data loaded by `Load()`. As such, its definition can be centralized via the `TopicRepositoryBase`. As part of this, the part of `Save()` that interacts with `GetContentTypeDescriptors()` was also able to be centralized as part of `TopicRepositoryBase`.