Implemented `GitVersionTask`
`GitVersionTask` is an *MSBuild* task that will automatically set the `Attribute*Version` properties in any compiled assemblies based on the *Git Flow* versioning system. In this model, the version will be based on a) the branch name, and/or b) previously set tags. The versions are expected to follow *Semantic Versioning* (SemVer).