Set `ConfigureAwait(false)` on all tasks
To avoid callers from shortcuting asynchronous tasks by e.g. calling `.Result` immediately, added `ConfigureAwait(false)` to all tasks, as per code analysis guidance. It's unclear if this is really needed on all downstream callers, or if it could just be added to e.g. `TopicController` and `TopicMappingService` (i.e., the root of the asynchronous methods being called in most cases). To alleviate warnings and ensure this is explicit, however, I've added this to all calls—even though it makes the code a bit more cumbersome and verbose.