Changed the default action to `Index`
Previously, the default action on the `Error` class was `Error`, thus yielding a route of `/Error/Error/`. That's dumb. I've changed this to `Index`, which is the default action for the route, thus yielding a route of `/Error/`. That makes more sense for the default (generic) error page. Also updated the unit tests to reflect this change.