Resolved bug in error tests
The tests for `NotFound` and `InternalServer` both called the default `Error` action, thus failing to test their use cases. In fixing them, this exposed a deeper issue since those actions depend on the `HttpContext` which is not set or available in tests (at least by default—it requires more effort to establish fakes for these). This is worked around here using a `null` check since the functionality isn't critical to the test.