Established copies of `(Cached)LayoutControllerBase{T}.cs`
Move the temporary copies of `(Cached)LayoutController{T}.cs` back to their original location. This, combined with the previous commits, effectively results in having a fork of these files—one in their original location, the other now in `(Cached)NavigationMappingService{T}.cs`. This allows the history of these files to be maintained even as the bulk of the business logic is moved to the `Ignia.Topics.Mapping` namespace as a service class that can be reused between both `Ignia.Topics.Web.Mvc` and `Ignia.Topics.AspNetCore.Mvc`.