Renamed to `Models`
There are other types of models, and not just view models, that *OnTopic* will need to account for. And, yet, there aren't likely enough of these to warrant differentiation at the level of `Ignia.Topics`. Further, the current name introduces ambiguity and potential conflict with the `Ignia.Topics.ViewModels` assembly, which utilizes the same namespace—and should continue to do so.