Upgraded to 4.8
Ideally, a class library should use the lowest necessary version in order to maintain compatibility with as broad of a selection of projects as possible. For .NET Standard 2.0 projects, that's technically 4.6.1—though Microsoft advises using 4.7.2 to ensure full compatibility. Since all of OnTopic's current implementers are on .NET Framework 4.8 or can easily upgrade to it, there shouldn't be any downside to targeting the latest framework—plus it permits full compatibility with the latest dependency updates.