Suppressed `CA1062` false positives
The `CA1062` warning pertains to *C# 8*'s new nullable annotation context. Unfortunately, variables that are ensured to not be null via `Contract.Requires()` or `Contract.Assure()` are not detected as such by Roslyn's flow analysis. We'll ultimately need to report this as a bug to the language team. In the meanwhile, they're distracting from actual issues by introducing false positives. I'm not a big fan of using global suppression files—_but_ this provides a clean way of coming back to these at a later date and removing them, once the core issue has been resolved.