Introduce project-specific solution file
Previously, this repository was part of the `Topics-Library` (soon to be `OnTopic-Library`) repository. As it has been split off, however, the previous `sln` file was left behind since it included a long commit history that wasn't specific to `OnTopic.Web.Mvc`. This fresh version is specific to `OnTopic.Web.Mvc`, and acknowledges that while this is legacy code, it is a new solution.