Moved `README` to root
When the `OnTopic-MVC` project was part of the `OnTopic-Library` project, the root `README.md` file was for the entirety of the OnTopic library. Now that the `OnTopic.Web.Mvc` project has been broken out into its own repository, the `OnTopic.Web.Mvc/README.md` makes sense in the root of the project. (There are other projects besides the `OnTopic.Web.Mvc` project, but they're closely related, like the `Host` and `Tests` projets.)