Updated to release version of `OnTopic-Library`
With the release today of **OnTopic 4.0.0**, we're able to set the dependencies to the latest release version. This will also support all versions above 4.0.0 in the future, in case implementors have a later version. As such, the library won't need to be kept up-to-date with each version of **OnTopic**, unless it needs to take advantage of a new feature.